+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * pga / golf is now available
 * atp/ tennis is now available
 * added function to pull a seasons results to team class / pyespn client
+* added im image class and connected to team logos and venue images
+* added in func to pull stats for a team/all teams of a league
+* added in call for coaches
 * 
 
 ## 0.3.0